requirements.txt
biopython
scipy
matplotlib
numpy